+++ lib/src/support/disposable.dart
@@ -1,10 +1,14 @@
+import 'dart:async';
+
+import 'package:flutter/foundation.dart';
 import 'package:meta/meta.dart';
 
 import '../extensions.dart';
 import '../logger.dart';
 
 typedef OnDisposeFunc = Future<void> Function();
-mixin Disposable {
+
+mixin _Disposer {
   //
   final _disposeFuncs = <OnDisposeFunc>[];
   bool _isDisposed = false;
@@ -13,8 +17,7 @@ mixin Disposable {
   // last added func will be called first when disposing
   void onDispose(OnDisposeFunc func) => _disposeFuncs.add(func);
 
-  @mustCallSuper
-  Future<bool> dispose() async {
+  Future<bool> _dispose() async {
     if (!_isDisposed) {
       logger.fine('[${objectId}] dispose()');
       _isDisposed = true;
@@ -32,5 +35,56 @@ mixin Disposable {
       logger.warning('[$objectId] unnecessary dispose() called.');
       return false;
     }
+  }
+}
+
+abstract class Disposable with _Disposer {
+  @mustCallSuper
+  Future<bool> dispose() async {
+    return await _dispose();
+  }
+}
+
+abstract class DisposableChangeNotifier extends ChangeNotifier with _Disposer {
+  @override
+  Future<bool> dispose() async {
+    if (!isDisposed) super.dispose();
+    return await super._dispose();
+  }
+
+  @override
+  bool get hasListeners {
+    if (isDisposed) {
+      logger.warning('called hasListeners on a disposed ChangeNotifier');
+      return false;
+    }
+    return super.hasListeners;
+  }
+
+  @override
+  void addListener(VoidCallback listener) {
+    if (isDisposed) {
+      logger.warning('called addListener() on a disposed ChangeNotifier');
+      return;
+    }
+    super.addListener(listener);
+  }
+
+  @override
+  void notifyListeners() {
+    if (isDisposed) {
+      logger.warning('called notifyListeners() on a disposed ChangeNotifier');
+      return;
+    }
+    super.notifyListeners();
+  }
+
+  @override
+  void removeListener(VoidCallback listener) {
+    if (isDisposed) {
+      logger.warning('called removeListener() on a disposed ChangeNotifier');
+      return;
+    }
+    super.removeListener(listener);
   }
 }
